Move ast_nodes_validators into appropriate module
